Replaced double to long double in lab 4.

+++ lab4/area.cpp
@@ -3,9 +3,9 @@
 
 #include "area.h"
 
-Area::Area(const Point<size_t> size, const double initial_approx) : size(size) {
+Area::Area(const Point<size_t> size, const long double initial_approx) : size(size) {
   size_t mem = (size.x + 2) * (size.y + 2) * (size.z + 2);
-  area = new double[mem];
+  area = new long double[mem];
   std::fill_n(area, mem, initial_approx);
 }
 
@@ -13,12 +13,12 @@ Area::~Area() {
   delete[] area;
 }
 
-double Area::get(const Point<int> pos) const noexcept {
+long double Area::get(const Point<int> pos) const noexcept {
   return area[pos.x * (size.y + 2) * (size.z + 2) + pos.y * (size.z + 2) +
               pos.z];
 }
 
-void Area::set(const double value, const Point<size_t> pos) noexcept {
+void Area::set(const long double value, const Point<size_t> pos) noexcept {
   area[pos.x * (size.y + 2) * (size.z + 2) + pos.y * (size.z + 2) + pos.z] =
       value;
 }
@@ -29,10 +29,10 @@ void Area::swapAreas(Area& area) noexcept {
 
 void Area::copyData(const Area& area) {
   size_t mem = (size.x + 2) * (size.y + 2) * (size.z + 2);
-  memcpy(this->area, area.area, mem * sizeof(double));
+  memcpy(this->area, area.area, mem * sizeof(long double));
 }
 
-double* Area::getFlatSlice(const size_t num) const {
+long double* Area::getFlatSlice(const size_t num) const {
   return area + (num * (size.y + 2) * (size.z + 2));
 }
